refactor: disassembled the product page into components
