+++ package.json
@@ -11,6 +11,7 @@
     "lint:scss": "npx stylelint \"**/*.scss\" --fix",
     "lint": "eslint ./src --ext ts,tsx --report-unused-disable-directives",
     "test": "jest --env=jsdom --coverage",
+    "format": "npm run prettier && npm run lint:scss && npm run lint",
     "prepare": "husky"
   },
   "devDependencies": {
