+++ src/pages/ProductPage/BlockFAQ/blockFAQ.module.scss
@@ -1,9 +1,20 @@
 .faqWrapper {
-  @include flex(none, none, column, 30px);
+  position: relative;
+
+  @include flex(center, center, column, 30px);
   z-index: 1;
 
   max-width: 100%;
-  margin-left: 35px
+  margin-left: 35px;
+
+  .plantImage {
+    position: absolute;
+    top: 0;
+    right: 0;
+    z-index: 0;
+
+    width: 380px;
+  }
 }
 
 .faqTitle {
@@ -11,22 +22,31 @@
 }
 
 .faqItemTitle {
-  @include montserrat(28px, 400, 30px);
-  width: 70%;
+  position: relative;
+  z-index: 1;
+
+  width: clamp(300px, 80vw, 1200px);
   padding: 10px 40px;
   border: 0.5px solid #898989;
   border-radius: 5px;
 
   text-align: start;
 
-  background: linear-gradient(.47turn, #3f3d3d 0%, $black 70%);
+  background: linear-gradient(.47turn, #3f3d3d90 0%, #17171695 70%);
   cursor: pointer;
+  backdrop-filter: blur(3px);
 
   user-select: none;
+
+  @include montserrat(28px, 400, 30px);
 }
 
 .faqItemAnswer {
-  @include montserrat(20px, 400, 35px);
-  width: 70%;
+  position: relative;
+  z-index: 1;
+
+  width: clamp(300px, 80vw, 1200px);
   margin-top: 0.5rem;
+
+  @include montserrat(20px, 400, 35px);
 }
