+++ src/pages/ProductPage/ProductSlider/productSlider.module.scss
