+++ src/pages/ProductPage/productCard/productCard.module.scss
@@ -1,0 +1,67 @@
+.productCard {
+  @include flex(null, null, row, 10px);
+  position: relative;
+
+  flex-wrap: wrap;
+
+  .imageCard {
+    margin-left: 35px;
+  }
+
+  .productData {
+    @include flex(null, null, column, 30px);
+    z-index: 1;
+
+    flex: 1 1 400px;
+    padding-left: 5px;
+
+    .productName {
+      @include montserrat(35px, 600, 40px);
+    }
+
+    .productDescription {
+      @include montserrat(16px, 400, 20px);
+    }
+
+    .price {
+      @include montserrat(30px, 600, 40px);
+      padding: 0 20px;
+      border: 1px solid white;
+      border-radius: 50px;
+    }
+
+    .crossed {
+      @include montserrat(20px, 600, 40px);
+      border: none;
+
+      color: $gray;
+      text-decoration: line-through;
+    }
+
+    .discountedPrice {
+      color: $red;
+    }
+  }
+
+}
+
+.plantImage {
+  position: absolute;
+  left: 0;
+  z-index: 0;
+}
+
+.buttonsWrapper,
+.priceWrapper {
+  @include flex(null, null, row, 30px);
+  justify-content: space-between;
+  width: 350px;
+}
+
+.buyButtons {
+  @include montserrat(20px, 600, 25px);
+  width: 160px;
+  height: 50px;
+
+  color: white;
+}
