+++ src/widgets/Footer/Footer.module.scss
@@ -1,5 +1,6 @@
 .socialMedia {
   position: relative;
+  z-index: 1;
 
   display: flex;
   flex-direction: column;
